pyproject.toml
[project]
name = "unify-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.10"
dependencies = [
    "discord>=2.3.2",
    "fastmcp>=2.11.1",
    "mcp>=1.12.3",
    "openai-agents>=0.2.4",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
